Website update
The files were renamed to their proper names. The font-faces were added for the custom font. The body and html background color was set. The navigation bar styles were set including the background color and when the viewer hovers over the button the background color changes. The notes and upcoming events section was styled using background colors and padding. The custom font was applied to the top h1. The beginning of the padding was made for the page.